backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
sse-starlette==2.2.1
python-multipart==0.0.20
aiofiles==24.1.0
google-genai==1.14.0
